Move tests to specific folders based on area
